in work
Осталось сделать список ошибок. Остановился на том, что получил список инвентаризации с ошибками дублирования.
