+++ Client_App/Views/MainWindow.axaml
@@ -120,10 +120,8 @@
                     Command="{Binding ExcelExportExecutors}"
                     CommandParameter="Executors"
                     Header="Список исполнителей"/>
-                <!--
                 <MenuItem Command="{Binding  ExcelExportCheckLastInventoryDate}"
                           Header="Проверка последней инвентаризации"/> 
-                -->
                 <MenuItem Header="Выбранная организация">
                     <MenuItem Header="Формы 1">
                         <MenuItem Command="{Binding ExcelExportForms}"
@@ -195,11 +193,12 @@
                     <MenuItem Command="{Binding ExcelExportAll}"
                               CommandParameter="SelectedOrg"
                               Header="Все формы"/>
-                    <!--
+                    
                     <MenuItem Header="СНК на дату">
                         <MenuItem Command="{Binding  ExcelExportSnk}"
                                   CommandParameter="1.1"
                                   Header="Форма 1.1"/>
+                        <!--
                         <MenuItem Command="{Binding  ExcelExportSnk}"
                                   CommandParameter="1.2"
                                   Header="Форма 1.2"/>
@@ -208,14 +207,13 @@
                                   Header="Форма 1.3"/>
                         <MenuItem Command="{Binding  ExcelExportSnk}"
                                   CommandParameter="1.5"
-                                  Header="Форма 1.5"/>
+                                  Header="Форма 1.5"/>-->
                     </MenuItem>
                     <MenuItem Header="Проверка инвентаризации">
                         <MenuItem Command="{Binding  ExcelExportCheckInventories}"
                                   CommandParameter="1.1"
                                   Header="Форма 1.1"/>
                     </MenuItem> 
-                    -->
                 </MenuItem>
                 <!--
                 <MenuItem Command="{Binding MaxGraphsLength, Mode=OneTime}"
